hw0
"You will write a C++ program to read records in a plain text file, then remove duplicated records, and write unique ones to a plain text file. You can use any algorithm and data structures that you prefer, as long as the results are correct. It is preferred, but not necessary, that your algorithm is as efficient as possible, both in processing time as well as memory management. The input is one text file with 0 - 10000 records. The content of each record is confined to a pair of {} and doesn’t contain ’{’ or ’}’, but may have other symbols like space, comma, colon, single or double quote, and so on. In the input and output file, each line ends in a ’\n’ character. One record may not be necessary in one line in the input file. But you should output each unique record in one line without any spaces. The output records can be in any orders. It doesn’t matter if you have or don’t have one empty line at the end of the output file."