Create hw2
A program that reads a text file containing equations that use long strings numbers and evaluates them using a stack, then returns the solution to each equation. Is capable of performing addition, subtraction, and multiplication. If an equation is unbalanced or invalid in any way, no solution will be given.