Create AVL Quiz
Program from a quiz meant to test our knowledge of AVL trees. Takes a text file containing a list of names and IDs, asks the user if they would like their AVL tree to be ordered according to ID or name, then creates an AVL tree based on user's response.