Create Stack Quiz - Palindrome Tester
A program used to test our skills when writing and implementing our own stacks. Uses two stacks made with linked lists to split given words from a text file in half, then check if each stack to see if their contents match. If so, the original word is printed in an output text file with the word "True" next to it. If contents don't match, the original word is printed to the output file with the word "False" next to it.