Create Equation Balance
An extra credit assignment for COSC 2430. Using a stack implemented as a linked list, this program reads through a string equation char by char until it finds an open or closed parenthesis, curly brace, or square bracket. When it finds one, finds the first open version of such a character, it pushes it into the stack. When the program finds the closing version of this character, it pops both from the stack. If after reading through the entire string equation the stack is empty, the equation is declared balanced. If not, it is imbalanced.