Create HW3
This program simulates a database with add, get, and deletion properties through the use of hash tables and binary search trees. The argumentmanager class is included because it was originally to be submitted through a linux server and that class was necessary for the specific way the user was expected to input commands.