Create GA5 - Gambling Reports
A group assignment using a circular linked list to add to a list of gamblers going to Las Vegas. Each person was given a randomly generated amount of starter money, which is reported in the initial report. In the final report, it is shown how many rounds each person played before the time limit was up, where each round was timed according to the Fibonacci sequence.